feat: Add utility modules for data collection workflows
- Introduced `collector_shared.py` for shared helpers in market data ingestion.
- Added `data_validation.py` for data validation utilities.
- Created `db_utils.py` for database connection and SQL statement utilities.
- Implemented `logging_utils.py` for error logging in the data collection pipeline.
- Developed `time_utils.py` for time manipulation and formatting utilities.
- Updated tests to reflect changes in module imports and functionality.
- Refactored existing tests to improve structure and maintainability.
