+++ .env.template
@@ -21,7 +21,7 @@ LOG_LEVEL=INFO
 LOG_DIR=./logs
 
 # Data Collection Settings
-SYMBOLS=AAPL,MSFT,GOOGL,AMZN
+SYMBOLS=AAPL,AMD,AMZN,BA,BABA,BAC,C,CSCO,CVX,DIS,F,GE,GOOGL,IBM,INTC,JNJ,JPM,KO,MCD,META,MSFT,NFLX,NVDA,PFE,T,TSLA,VZ,WMT,XOM
 COLLECTION_INTERVAL_MINUTES=60
 WINDOW_OVERLAP_MINUTES=5
 COLLECTION_SCHEDULE=*/5 * * * *  # Every 5 minutes
