+++ src/model/__init__.py
@@ -1,0 +1,1 @@
+"""ORM models and database session helpers."""
